Three passes successive removal data of A. salmon in the Nivelle river

A dataset containing the successive removal data electofishing of salmon juveniles over 3 passes on 3 years (2003--2005) and 11 sites. Contains missing data.

Usage

SucRemNivelle

Format

A named list:

I

The number of sites (sites x years)

C1

The number of captures of the first pass for each site

C2

The number of captures of the second pass for each site

C3

The number of captures of the third pass for each site

S

The area in squared meter that was fished for each site

AI

The abundance index of each site

I_pred

The number of sites (for prediction)

AI_pred

Abundance indices (for prediction)

Source

Brun, M., Abraham, C., Jarry, M., Dumas, J., Lange, F., & Prévost, E. (2011). Estimating an homogeneous series of a population abundance indicator despite changes in data collection procedure: A hierarchical Bayesian modelling approach. Ecological Modelling, 222(5), 1069-1079.